backend/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
Werkzeug==0.16.1
zipp==3.1.0
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
Werkzeug==0.16.1
zipp==3.1.0
